requirements.txt
python-telegram-bot>=21.0
msal>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
